Split Database.js into feature-related files
